# task-manager-backend/src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task Manager API",
    "version": "1.0.0",
    "description": "API para gestionar tareas"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor local"
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Registrar un nuevo usuario",
        "description": "Crea un nuevo usuario en el sistema",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "lastname": { "type": "string" },
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              },
              "example": {
                "name": "Juan",
                "lastname": "Pérez",
                "email": "juan@example.com",
                "password": "123456"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado exitosamente",
            "content": {
              "application/json": {
                "example": { "id": 1, "name": "Juan", "lastname": "Pérez", "email": "juan@example.com" }
              }
            }
          },
          "400": {
            "description": "Email ya registrado"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Iniciar sesión",
        "description": "Autentica un usuario y retorna un token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              },
              "example": {
                "email": "juan@example.com",
                "password": "123456"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login exitoso",
            "content": {
              "application/json": {
                "example": { "id": 1, "name": "Juan", "lastname": "Pérez", "email": "juan@example.com", "token": "uuid-token" }
              }
            }
          },
          "401": {
            "description": "Credenciales inválidas"
          }
        }
      }
    },
    "/task": {
      "get": {
        "summary": "Listar todas las tareas",
        "description": "Retorna todas las tareas registradas",
        "responses": {
          "200": {
            "description": "Lista de tareas",
            "content": {
              "application/json": {
                "example": [
                  { "id": 1, "title": "Tarea 1", "description": "Descripción 1", "created_at": "2024-01-01T00:00:00.000Z" }
                ]
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear una tarea",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" }
                }
              },
              "example": {
                "title": "Nueva tarea",
                "description": "Descripción de la tarea"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tarea creada"
          },
          "400": {
            "description": "Error de validación"
          }
        }
      }
    },
    "/task/{id}": {
      "get": {
        "summary": "Obtener detalle de una tarea",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle de la tarea"
          },
          "404": {
            "description": "Tarea no encontrada"
          }
        }
      },
      "put": {
        "summary": "Actualizar una tarea",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tarea actualizada"
          },
          "404": {
            "description": "Tarea no encontrada"
          }
        }
      },
      "delete": {
        "summary": "Eliminar una tarea",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "204": {
            "description": "Tarea eliminada"
          },
          "404": {
            "description": "Tarea no encontrada"
          }
        }
      }
    }
  }
}